+++ crates/swc/tests/fixture/issues-5xxx/5752/output/index.ts
@@ -1,0 +1,4 @@
+var foo = {
+    bar: 123
+};
+delete foo.bar;