+++ crates/swc/tests/tsc-references/assertionTypePredicates2.1.normal.js
@@ -7,7 +7,7 @@
  * @param {A} a
  * @returns { asserts a is B }
  */ var foo = function(a) {
-    if (/** @type { B } */ (a).y !== 0) throw TypeError();
+    if (/** @type { B } */ a.y !== 0) throw TypeError();
     return undefined;
 };
 export var main = function() {