+++ crates/swc/tests/tsc-references/classWithStaticFieldInParameterBindingPattern(target=es5).1.normal.js
@@ -1,7 +1,7 @@
 //// [classWithStaticFieldInParameterBindingPattern.ts]
-// https://github.com/microsoft/TypeScript/issues/36295
 import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
 var _class;
+// https://github.com/microsoft/TypeScript/issues/36295
 (function(param) {
     var tmp = param[(_class = function _class() {
         "use strict";