+++ crates/swc/tests/tsc-references/typeGuardsInRightOperandOfAndAndOperator.1.normal.js
@@ -29,14 +29,14 @@ function foo6(x) {
     return typeof x !== "string" // string | number | boolean
      && (typeof x !== "number" // number | boolean
      ? x // boolean
-     : x === 10) // number 
-    ;
+     : x === 10 // number 
+    );
 }
 function foo7(x) {
     var y;
     var z;
     // Mixing typeguard narrowing
-    return typeof x !== "string" && (z = x) // number | boolean
-     && (typeof x === "number" ? (x = 10) && x.toString() // x is number
+    return typeof x !== "string" && (z = x // number | boolean
+    ) && (typeof x === "number" ? (x = 10) && x.toString() // x is number
      : (y = x) && x.toString()); // x is boolean
 }