+++ crates/swc_node_comments/src/lib.rs
@@ -36,10 +36,14 @@ impl Comments for SwcComments {
     }
 
     fn move_leading(&self, from: BytePos, to: BytePos) {
-        let cmt = self.leading.remove(&from);
+        let cmt = self.take_leading(from);
 
-        if let Some(cmt) = cmt {
-            self.leading.entry(to).or_default().extend(cmt.1);
+        if let Some(mut cmt) = cmt {
+            if from < to && self.has_leading(to) {
+                cmt.extend(self.take_leading(to).unwrap());
+            }
+
+            self.add_leading_comments(to, cmt);
         }
     }
 
@@ -68,10 +72,14 @@ impl Comments for SwcComments {
     }
 
     fn move_trailing(&self, from: BytePos, to: BytePos) {
-        let cmt = self.trailing.remove(&from);
+        let cmt = self.take_trailing(from);
 
-        if let Some(cmt) = cmt {
-            self.trailing.entry(to).or_default().extend(cmt.1);
+        if let Some(mut cmt) = cmt {
+            if from < to && self.has_trailing(to) {
+                cmt.extend(self.take_trailing(to).unwrap());
+            }
+
+            self.add_trailing_comments(to, cmt);
         }
     }